py.test conftest fixes, details see below, updated copyright
init_test_app / deinit_test_app: removed the calls on class level
this should not be needed as we call them with setup_method and
teardown_method and we don't have support for nesting these calls.

init_test_app on module level: removed, for similar reasons.
I checked the tests with and without this and we seem to no need init_test_app
on module level. There was no call to deinit_test_app on module level btw.
In case we ever need this, we could maybe do it on function level (not module
level) using setup_function and teardown_function.

teardown_method: changed order of calls: we need to first call the wrapped
classes' teardown_method and afterwards deinit the test app. Otherwise it
crashes if the classes' teardown_method uses stuff removed by deinit.