AccessControlList: do not require cfg object
Giving the cfg to the ACL object caused some issues (for example, the AclTokenizer
in the whoosh branch needed cfg and was part of the schema. whoosh tries to
pickle the schema and that did not work for cfg).

Refactored "auth_trusted" - this is now a user object attribute, so the ACL
code does not need to know the list of trusted auth methods (which is still in
cfg, but it shouldn't even be there).

It is AccessControlList(acl_lines=[], default='', valid=None) now, you need
to give the list of valid rights strings via valid param.

TODO: deal with cfg.acl_rights_functions, cfg.acl_rights_contents (defaults?)

Got rid of ContentACL and FunctionACL classes, they were pointless as "valid"
needs to be given anyway.