use main part of the namespace for user items, use uuid4 for user item uuids
we use the normal wiki namespace for now to store user profile items, using
the user name as item name. for now, this is simpler. if we feel the need,
we can go back to storing them into UserProfile/* part of the namespace.

If we do that, there might be the issue that itemname != username then.