fixed the routing middleware
the routing mw iterator now yields tuples (mountpoint, revid), some
methods in the routing middleware want name,revid now (this is so
they can find out the correct backend, with just the revid this is not
easily possible (would need an index lookup for the name).

the indexing tests now just use the storage given by the test framework
instead of manually assembling their own one.

most changes in this cs are related to these changes.

additionally I introduced some new test class attribute "reinit_storage" -
if it is present and True, the test framework will recreate the wsgi app
(and storage and index) for every test method call in that class.