Better output for maint_reduce_revisions command.
For each item, it prints out the current name and revids of all historical
revisions, each in an indented line.  When the item name of a historical
revision is different from current name, the historical name is printed too.
This makes the message clearer, and unambiguous in case there are cross
renamings.